requirements.txt
# Core Dependencies
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
python-dotenv==1.0.0

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.21
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication & Security
Flask-Login==0.6.2
Flask-WTF==1.2.1
email-validator==2.1.0.post1
PyJWT==2.8.0
Flask-Talisman==1.0.0
Flask-Cors==4.0.0
Flask-Limiter==3.5.0

# Firebase
firebase-admin==6.2.0
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0

# Google AI
google-generativeai==0.3.2

# Email
Flask-Mail==0.9.1

# File Processing
Pillow==10.0.1
python-magic==0.4.27
python-magic-bin==0.4.14; sys_platform == 'win32'

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
python-slugify==8.0.1
bleach==6.0.0

# Development
debugpy==1.6.7
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.11.1
coverage==7.3.1
black==23.9.1
flake8==6.1.0
isort==5.12.0
mypy==1.5.1
types-python-dateutil==2.8.19.13
types-requests==2.31.0.1

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Production
gunicorn==21.2.0
gevent==23.7.0

# Local Development
pyngrok==7.0.0